+++ build-logic/library-convention/src/main/kotlin/com/cvix/buildlogic/library/LibraryConventionPlugin.kt
@@ -23,7 +23,6 @@ class LibraryConventionPlugin : ConventionPlugin {
         dependencies {
 
             add("implementation", catalogBundle("kotlin-jvm"))
-            add("implementation", catalogBundle("jackson"))
         }
     }
 }