+++ build-logic/spring-boot-convention/src/main/kotlin/com/cvix/buildlogic/springboot/SpringBootConventionPlugin.kt
@@ -19,7 +19,6 @@ class SpringBootConventionPlugin : ConventionPlugin {
         apply(plugin = catalogPlugin("kotlin-jvm").get().pluginId)
         apply(plugin = catalogPlugin("kotlin-spring").get().pluginId)
         apply(plugin = catalogPlugin("spring-boot").get().pluginId)
-        apply(plugin = catalogPlugin("spring-dependency-management").get().pluginId)
 
         with(extensions) {
             commonExtensions()
@@ -31,7 +30,6 @@ class SpringBootConventionPlugin : ConventionPlugin {
             add(IMPLEMENTATION, catalogBundle("spring-boot"))
 
             add(IMPLEMENTATION, catalogBundle("kotlin-jvm"))
-            add(IMPLEMENTATION, catalogBundle("jackson"))
 
             add("developmentOnly", catalogBundle("spring-boot-dev"))