+++ server/engine/src/main/kotlin/com/cvix/authentication/infrastructure/SecurityConfiguration.kt
@@ -152,7 +152,8 @@ class SecurityConfiguration(
             .redirectToHttps {
                     httpsRedirect ->
                 httpsRedirect.httpsRedirectWhen {
-                    it.request.headers.containsKey("X-Forwarded-Proto")
+                    val forwardedProto = it.request.headers.getFirst(X_FORWARDED_PROTO)?.lowercase()
+                    forwardedProto != null && forwardedProto != "https"
                 }
             }
             .headers { headers ->
@@ -205,6 +206,7 @@ class SecurityConfiguration(
                 "/v3/api-docs/**", "/v3/api-docs.yaml",
             ).permitAll()
             .pathMatchers(HttpMethod.POST, "/api/waitlist").permitAll()
+            .pathMatchers(HttpMethod.POST, "/api/contact").permitAll()
             .pathMatchers("/actuator/**").authenticated()
             .pathMatchers("/api/**").authenticated()
             .pathMatchers("/management/health").permitAll()