+++ shared/spring-boot-common/src/main/kotlin/com/cvix/spring/boot/presentation/filter/RHSFilterParserFactory.kt
@@ -1,9 +1,9 @@
 package com.cvix.spring.boot.presentation.filter
 
 import com.cvix.common.domain.presentation.filter.RHSFilterParser
-import com.fasterxml.jackson.databind.ObjectMapper
 import kotlin.reflect.KClass
 import org.springframework.stereotype.Component
+import tools.jackson.databind.ObjectMapper
 
 @Component
 class RHSFilterParserFactory(