+++ shared/spring-boot-common/src/main/kotlin/com/cvix/spring/boot/presentation/sort/SortParserFactory.kt
@@ -1,8 +1,8 @@
 package com.cvix.spring.boot.presentation.sort
 
-import com.fasterxml.jackson.databind.ObjectMapper
 import kotlin.reflect.KClass
 import org.springframework.stereotype.Component
+import tools.jackson.databind.ObjectMapper
 
 @Component
 class SortParserFactory(